优化查询商品模块功能

+++ mall_common/src/main/java/com/mall/common/pojo/EasyUIDataGridResult.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 import java.util.List;
 
 /**
+ * 查询商品返回类型
  * @author Sam
  * @date 2018/9/18
  * @time 20:49
@@ -11,14 +12,14 @@ import java.util.List;
 public class EasyUIDataGridResult implements Serializable {
     private static final long serialVersionUID = -6849721563764815745L;
 
-    private Integer total;
+    private Long total;
     private List<?> rows;
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
